[Feat] MainButton 스타일 재배치, 디자인에 맞게 수정
