+++ janghakhere/janghakhere/Feature/Onboarding/MainButtonView.swift
@@ -32,6 +32,22 @@ struct MainButtonView: View {
     }
 }
 
+struct MainButtonStyle: ButtonStyle {
+  func makeBody(configuration: Configuration) -> some View {
+    configuration.label
+      .font(.title_xsm)
+      .padding(.vertical, 16)
+      .foregroundStyle(.white)
+      .frame(maxWidth: .infinity)
+      .frame(height: 58)
+      .background(
+        Capsule()
+          .fill(.mainGray)
+      )
+      .scaleEffect(configuration.isPressed ? 0.95 : 1.0)
+  }
+}
+
 #Preview {
     MainButtonView(title: "다음", clickedButton: {})
 }
