+++ janghakhere/janghakhere/Model/Onboarding/Gender.swift
@@ -1,0 +1,14 @@
+//
+//  Gender.swift
+//  janghakhere
+//
+//  Created by Taehwan Kim on 4/21/24.
+//
+
+import Foundation
+
+enum Gender {
+    case notSelected
+    case female
+    case male
+}
